CPM-541: refactor integration test
